+++ src/models/Settings.php
@@ -25,13 +25,68 @@ class Settings extends Model
     // Public Properties
     // =========================================================================
 
-    /**
-     * @var string
-     */
-    public $someAttribute = 'Some Default';
+    /** @var string */
+    public $ipapiKey = null;
+
+    /** @var boolean */
+    public $gdprBanner = true;
+
+    /** @var string */
+    public $impliedHeading;
+
+    /** @var string */
+    public $impliedCopy;
+
+    /** @var string */
+    public $explicitHeading;
+
+    /** @var string */
+    public $explicitCopy;
+
+    /** @var string */
+    public $defaultConsentType = 'none';
+
+    /** @var array */
+    public $consentTypes;
+
+    /** @var array */
+    public $impledConsentTable = [];
+
+    /** @var array */
+    public $explicitConsentTable = [];
 
     // Public Methods
     // =========================================================================
+
+    public function init()
+    {
+        $this->consentTypes = [
+            [
+                'handle' => 'necessary',
+                'name' => Craft::t('total-cookie-consent', 'Necessary'),
+                'defaultOn' => true,
+                'required' => true,
+            ],
+            [
+                'handle' => 'statistics',
+                'name' => Craft::t('total-cookie-consent', 'Statistics'),
+                'defaultOn' => true,
+                'required' => false,
+            ],
+            [
+                'handle' => 'marketing',
+                'name' => Craft::t('total-cookie-consent', 'Marketing'),
+                'defaultOn' => true,
+                'required' => false,
+            ],
+        ];
+
+        $this->explicitHeading = Craft::t('total-cookie-consent', 'This website uses cookies');
+        $this->explicitCopy = Craft::t('total-cookie-consent', "We use cookies to personalize content and ads, to provide social media features, and to analyze our traffic. We also share information about your use of our site with our social media, advertising and analytics partners who may combine it with other information that you've provided to them or that they've collected from your use of their services.");
+
+        $this->impliedHeading = Craft::t('total-cookie-consent', 'This website uses cookies');
+        $this->impliedCopy = Craft::t('total-cookie-consent', "We use cookies to personalize content and ads, to provide social media features, and to analyze our traffic. We also share information about your use of our site with our social media, advertising and analytics partners who may combine it with other information that you've provided to them or that they've collected from your use of their services.");
+    }
 
     /**
      * @inheritdoc
@@ -39,8 +94,10 @@ class Settings extends Model
     public function rules()
     {
         return [
-            ['someAttribute', 'string'],
-            ['someAttribute', 'default', 'value' => 'Some Default'],
+            [['impliedHeading', 'impliedCopy', 'explicitHeading', 'explicitCopy', 'consentTypes'], 'required'],
+            [['ipapiKey', 'impliedHeading', 'impliedCopy', 'explicitHeading', 'explicitCopy'], 'string'],
+            [['ipapiKey'], 'boolean'],
+            [['defaultConsentType'], 'in', 'range' => ['none', 'implied', 'explicit']],
         ];
     }
 }